rename camoComponent
