+++ src/components/CamouflageComponent.vue
